stealth-startup/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
boto3==1.35.19
botocore==1.35.19
certifi==2024.8.30
charset-normalizer==3.3.2
cohere==5.9.2
distro==1.9.0
exceptiongroup==1.2.2
fastavro==1.9.7
filelock==3.16.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
groq==0.11.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.7
idna==3.9
jiter==0.5.0
jmespath==1.0.1
openai==1.45.0
packaging==24.1
parameterized==0.9.0
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
slack_sdk==3.32.0
smmap==5.0.1
sniffio==1.3.1
tokenizers==0.20.0
tqdm==4.66.5
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
